requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
python-pptx>=1.0
pytest>=7.0
nbconvert>=7.0
ipykernel>=6.0
